+++ src/core/zowe/core_for_zowe_sdk/profile_manager.py
@@ -338,7 +338,8 @@ class ProfileManager:
         profile_type: Optional[str] = None,
         check_missing_props: bool = True,
         validate_schema: Optional[bool] = True,
-        override_with_env: Optional[bool] = False
+        validate_only_project_config: Optional[bool] = False,
+        override_with_env: Optional[bool] = False,
     ) -> dict[str, Any]:
         """
         Load connection details from a team config profile.
@@ -363,6 +364,8 @@ class ProfileManager:
             Flag to indicate whether to check for missing secure properties.
         validate_schema : Optional[bool]
             Whether to validate the loaded profile against the schema defined in the configuration.
+        validate_only_project_config : Optional[bool]
+            Whether to only load from project-level configs (project and project user), ignoring global configs.
         override_with_env : Optional[bool]
             If True, overrides profile properties with values from environment variables.
 
@@ -405,10 +408,20 @@ class ProfileManager:
             self.__global_config,
         ):
             if cfg_layer.profiles is None:
+
+                layer_validate_schema = validate_schema
+
+                if validate_only_project_config:
+                    layer_validate_schema = cfg_layer == self.__project_config
+
                 try:
-                    cfg_layer.init_from_file(validate_schema)
+                    cfg_layer.init_from_file(layer_validate_schema)
+
                 except SecureProfileLoadFailed:
-                    self.__logger.warning(f"Could not load secure properties for {cfg_layer.filepath}")
+                    self.__logger.warning(
+                        f"Could not load secure properties for {cfg_layer.filepath}"
+                    )
+
                     warnings.warn(
                         f"Could not load secure properties for {cfg_layer.filepath}",
                         SecurePropsNotFoundWarning,
@@ -452,7 +465,13 @@ class ProfileManager:
 
         if profile_type != BASE_PROFILE:
             profile_props = {
-                **self.load(profile_type=BASE_PROFILE, check_missing_props=False),
+                **self.load(
+                    profile_type=BASE_PROFILE,
+                    check_missing_props=False,
+                    validate_schema=validate_schema,
+                    validate_only_project_config=validate_only_project_config,
+                    override_with_env=override_with_env,
+                ),
                 **profile_props,
             }